Edits to FluxSolver implementation
The FluxSolver implementation previously made use of the controversial
'diamond' inheritance problem. This meant that each version of the
FluxSolver had to inherit from two parents -- one of which provided
the linear reconstruction (or lack thereof), and one which provided
a simple two-cell flux calculator. Each combination had to be added
to the code by hand and given a unique name. While manageable at
this level, this could become unworkable with the addition of a few
more calculators and a few more reconstruction methods. The model
was simplified such that a FluxSolver has a calculator and a
reconstructor, rather than is a calculator and a reconstructor.

Note that older settings files have not been updated to account for
these changes. See the file MaxwellCutCell.cfg for the new FVMsolver
format.

Other changes:
* Changed file reading protocol, now need to specify full path and
  any suffixes. This allows greater flexibility in config file
  naming conventions.
* Added some exception handling to libconfig reads. If some params
  aren't specified, they may now be caught and replaced with a
  default setting.
* Renamed LimiterNone to LimiterDefault. Hoping to role out similar
  changes throughout the code.
* Some directory reshuffling.