Removed Updater class
The Updater class has been deprecated as it did not provide any
useful functionality. It's primary purpose was to control
dimensional splitting, but the splitting scheme was hard-coded.
This functionality will be provided instead by the upcoming
EventSequencer, which will generalise this process.
The main 'Solver' class in FVlite.hpp (name may change in future
builds) has been renamed to 'Controller', since 'Solver' will
refer to a different set of classes in future. The functionality
of Updater has been moved here.