Solver overhaul -- Decorator Pattern
Solvers have now been overhauled to make use of the decorator
pattern (or more accurately, a slight variation on it). This
allows the user to choose their own series of events at run time.

This replaces the EventSequencer class, which was a work-in-
progress at the time of this update.

This change does delegate a lot of responsibility to the end-user,
so a collection of commonly used event sequences will be
implemented at a later date.

This change also included significant edits to the BoundaryMan-
ager class and the awful GridManager class.