Implemented ExplicitUpdater
A new class, ExplicitUpdater, has been introduced. It is used by
SolverFiniteVolume1D to perform grid updates, and currently
has two variants: Euler (standard FV update) and CutCells, which
implements the modified update formula for Klein cut cels.

SolverCutCellNewCycle has also been introduced, and this now
bears the responsibility of preparing the grid for cut cell
operations each time step. This functionality was removed from
CutCellManager, as this class is now owned exclusively by
ExplicitUpdaterCutCells, as it would be inappropriate to pass
'new time step' messages all the way from the controller
to the ExplicitUpdater, and then to have the ExplicitUpdater
perform a role unrelated to its primary function.

SolverFiniteVolume1DCutCell has been removed to account for these
changes. This will allow for the creation of composite Solvers,
such as SolverStrangSplitting, without having to specify whether
or not the splitting method should include cut cells or not.