WIP -- Solver to Operator overhaul
Not a functional build. In the process of replacing the
awful 'GridManager' class with a far more general
'Composite' class. Replacing 'Solvers' with 'Operators',
and using the composite design pattern to build multi-stage
solvers rather than the decorator pattern.

This change has resulted in code refactoring in many
places, including boundary management and initialisation.