New Branch - Visitor Experiment
New branch created to accomodate implementation of generic
visitor pattern.

Branched from main development as many design issues were not
foreseen, and it is not clear if the visitor pattern is really
appropriate. Double-dispatch is not a requirement, so the worry
is that the code will become unnecessarily complicated and that
the visitor will only serve as a maintenance nightmare and a
'God Object'. Seems like a lot of trouble just to ensure that
Operators no longer require a common interface.