Grid Overhaul Complete
The Grid has been overhauled in design in order to make it more
modular. Previously, if a new feature was to be added, it had to
be shoehorned into the existing Grid class. This led to it becoming
quite bloated, and the problem was due to get much worse.
Additionally, the implementation of some features was quite
frustrating to deal with (LevelSet was terrible!).

A modular Cartesian Cell-Centred Grid was implemented using a
variadic template design -- with more work, it could be generalised
further such that it would not necessarily be Cartesian Cell-Centred.
This design allows you to specify the dimensions of the Grid and
what is stored at each Grid location by specifying template
parameters. The working Grid is then included in the code as
follows:

typedef CartesianGridImplementation<
    DIM,
    StateGridCartesianCentred<DIM>,
    FluxGridCartesianCentred<DIM>,
    LevelSetGridCartesianCentred<DIM>,
    CutCellBoundaryStateGrid<DIM>,
    CutCellGeometryGrid<DIM>
> Grid;

The contents of the grid can be easily modified to adding or
removing subgrids from this parameter list. Each subgrid is
a simple implementation which bundles a std::vector (or perhaps
several) with access functions. Indexing and determining the size
of the grid is provided by a base class. All subgrids have access
to a single instance of this base class by virtual inheritance.
This way, all have the same size and are indexed the same way.

TODO:
-> Add EM materials subgrid, update flux calculators to account
   for this.
-> Decouple object construction from the objects themselves.
   It's annoying to require libconfig code everywhere, and the
   Controller class should not need to know how to do this.
-> Initialiser and Boundary overhaul.
-> Output overhaul.
-> Remove requirement that all Operators need common interface.
   Use Visitor + Composite (+ Decorator) + Iterator.