Euler equation fixes
All recent updates have concerned the Maxwell solvers, and little
attention was paid to the Euler equations. A few files have been
updated and a new test case added to account for this.

Note that the new test case fails when MUSCL reconstruction is
applied. Some quick tests suggest that this occurs regardless
of limiter choice. May be worth investigating this.