In Progress: Materials Implementation
WARNING: Euler currently broken.
         MaxwellCutCells broken.

Working on a 'Materials' class which encapsulates physical
features that are not part of a conservative state vector.
Currently only implemented for Maxwell's equations, and
hardcoded in some areas. Richtmyer flux (and therefore
FORCE) has a bias for leftmost material, which is not really
acceptable. Lax-Friedrichs works, and gives reasonable results
for test with a dielectric cylinder.

Planning to implement a version for Euler equations that
gives the appearance of being a grid while actually being
a single double (which should be gamma). Will move a lot
of functionality out of StateVector, and will overhaul the
mechanism for switching between Maxwell and Euler.
(Will probably use typedefs).