Further Grid Upgrades
Following a previous commit in which Grids were implemented in a
semi-generic fashion, further upgrades were made to increase the
modularity of the system:
    > BaseGridCartesianCentred replaced with two classes:
      BaseCartesianGrid, and CartesianCellCentredGrid. The former
      implements such information as number of cells, cell
      spacings, etc, while the latter implements the total number
      of elements (equal to number of cells in this case, but not
      true of other Cartesian grids such as vertex centred).
    > Generic CartesianGrid replaces GridImplementation. Stored
      alongside the base class.
    > SubGrids now generic. StateGridCartesianCentred<dim> is now
      StateGrid< CartesianCellCentredGrid<dim> >.
    > ConstantGrid implemented, which mimics a grid but does not
      store values at each location. This allows the classes
      MaterialGridCartesianCentred and MaterialGridConstant to be
      merged. Each case can be accessed as:
      MaterialGrid< Material, CartesianCellCentredGrid<dim> >
      MaterialGrid< Material, ConstantGrid >
These changes are designed to allow different types of Cartesian
grid to be implemented, such as edge cented, face centred, and
vertex centred Grids. A future update may result in level sets
being stored on the vertices rather than the cell centres, and
the implementation of the Yee grid will require edge centred and
face centred grids specialised to each dimension.