Specialised start/end/position functions
Functions to get start and end of grids in each dimension
are now specialised to the type of Cartesian grid. Calling
them may be a hassle, however. If all components of a grid
are cell centred, it should be sufficient to call:

int start_x = grid.start( DIM_X );

However, if multiple types of grid are included, we require:

int start_x = grid.CartesianCellCentredGrid<MY_DIM>::start( DIM_X );

This leads to fairly unreadable code. This change may be
reverted, since it may be easier to simply do:

int start_x = grid.end( DIM_X ) + 1;

in the cases where there is one extra element in the x direction.
It's not a pleasant convention, but it's easier to remember than:

int start_x  =  \
    grid.CartesianFaceCentredGrid<MY_DIM,DIM_X>::start( DIM_X);

Position finding functions also moved to specific Cartesian grid
types, since it is currently assumed that the user is interested
in finding cell-center positions. This leads to lots of repeated
code in which vertex positions are derived from cell-centers, but
this shouldn't be a manual operation.